refactor(async-agents): extract waitForResultFiles helper, exclude fire-and-forget from wait
